Move money flow services into MoneyFlow feature folders #388
InvestmentPaycheck is used instead of the suggested InvestmentRate
folder name because that segment would shadow the Domain
InvestmentRate entity inside the feature.
