+++ code/FinanceManager.Api/Controllers/Accounts/CurrencyAccountController.cs
@@ -2,7 +2,6 @@ using FinanceManager.Api.Helpers;
 using FinanceManager.Application.FinancialAccounts.Currencies;
 using FinanceManager.Application.FinancialAccounts.Shared.Exports;
 using FinanceManager.Application.Services;
-using FinanceManager.Application.Services.Currencies;
 using FinanceManager.Domain.Commands.Account;
 using FinanceManager.Domain.Dtos;
 using FinanceManager.Domain.Entities.Exports;
