+++ code/FinanceManager.Api/Controllers/Accounts/CurrencyAccountImportController.cs
@@ -3,7 +3,6 @@ using FinanceManager.Api.Helpers;
 using FinanceManager.Api.Hubs;
 using FinanceManager.Api.Services;
 using FinanceManager.Application.FinancialAccounts.Currencies.Import;
-using FinanceManager.Application.Services.Currencies;
 using FinanceManager.Domain.Dtos;
 using FinanceManager.Domain.Entities.FinancialAccounts.Currencies;
 using FinanceManager.Domain.Entities.Imports;
