+++ code/FinanceManager.Api/Services/CurrencyImportBackgroundService.cs
@@ -1,7 +1,6 @@
 using FinanceManager.Api.Hubs;
 using FinanceManager.Application.Diagnostics;
 using FinanceManager.Application.FinancialAccounts.Currencies.Import;
-using FinanceManager.Application.Services.Currencies;
 using FinanceManager.Domain.Entities.Imports;
 using Microsoft.AspNetCore.SignalR;
 using System.Diagnostics;
