+++ code/FinanceManager.Tests.Unit/Application/Services/LiabilitiesServiceTests.cs
@@ -1,3 +1,4 @@
+using FinanceManager.Application.MoneyFlow.NetWorth;
 using FinanceManager.Application.Services;
 using FinanceManager.Domain.Entities.FinancialAccounts.Currencies;
 using FinanceManager.Domain.Enums;
